pyproject.toml
[project]
name = "nonebot-plugin-brick"
version = "0.1.0"
description = "群聊娱乐插件：烧砖拍人，一砖撂倒（移植自 koishi-plugin-brick）"
authors = [
    {name = "SZ2528",email = "83620631+SZ2528@users.noreply.github.com"}
]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "nonebot2>=2.4.3",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
